adding kembalian on penjualan

+++ database/migrations/2022_12_15_133124_create_pembelian_detail_table.php
@@ -17,7 +17,7 @@ class CreatePembelianDetailTable extends Migration
             $table->bigIncrements('id_pembelian_detail');
             $table->integer('id_pembelian');
             $table->integer('id_produk');
-            $table->integer('qty');
+            $table->float('qty', 5, 2);
             $table->integer('harga');
             $table->integer('subtotal');
             $table->string('created_by');
